Main rendering function finished
I finished the main render function and added some comments.

+++ Rally-X/Player.cpp
@@ -103,7 +103,7 @@ void Player::update(int type)
 	}
 }
 
-int Player::getFuel()
+double Player::getFuel()
 {
 	return fuelLevel;
 }
